# PharmacyOnline/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "PharmacyOnline, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Product/admin/create/product/tablet": {
      "post": {
        "tags": [
          "Product"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "cateId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "productName",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "title",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "thumbnail",
            "type": "file"
          },
          {
            "in": "formData",
            "name": "modelNumber",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "dies",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "maxPressure",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "maxDiameterOfTablet",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "maxDepthOfFill",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "productionCapacity",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "machineSize",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "netWeight",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Product/admin/create/product/capsule": {
      "post": {
        "tags": [
          "Product"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "cateId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "productName",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "title",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "thumbnail",
            "type": "file"
          },
          {
            "in": "formData",
            "name": "Output",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "CapsuleSize",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "MachineDimension",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "ShippingWeight",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Product/admin/create/product/liquid": {
      "post": {
        "tags": [
          "Product"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "cateId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "productName",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "title",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "thumbnail",
            "type": "file"
          },
          {
            "in": "formData",
            "name": "AirPressure",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "AirVolume",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "FillingSpeed",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "FillingRange",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Product/admin/update/product/tablet": {
      "post": {
        "tags": [
          "Product"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "productName",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "title",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "thumbnail",
            "type": "file"
          },
          {
            "in": "formData",
            "name": "modelNumber",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "dies",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "maxPressure",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "maxDiameterOfTablet",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "maxDepthOfFill",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "productionCapacity",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "machineSize",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "netWeight",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Product/admin/update/product/capsule": {
      "post": {
        "tags": [
          "Product"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "productName",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "title",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "thumbnail",
            "type": "file"
          },
          {
            "in": "formData",
            "name": "Output",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "CapsuleSize",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "MachineDimension",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "ShippingWeight",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Product/admin/update/product/liquid": {
      "post": {
        "tags": [
          "Product"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "productName",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "title",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "thumbnail",
            "type": "file"
          },
          {
            "in": "formData",
            "name": "AirPressure",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "AirVolume",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "FillingSpeed",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "FillingRange",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Product/getdetail": {
      "get": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idProduct",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Product/admin/delete": {
      "get": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "productId",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Product": {
      "get": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "isNewest",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "search",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "cate",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "sorting",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32",
            "default": 1,
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
...
}